python-bridge/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
pyzk==0.9
pyserial==3.5
python-multipart==0.0.6
